+++ src/main/java/subway/line/LineController.java
@@ -1,13 +1,8 @@
 package subway.line;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.section.Section;
-import subway.section.SectionDao;
-import subway.section.SectionRequest;
-import subway.section.SectionService;
 import subway.station.StationResponse;
 
 import java.net.URI;
@@ -16,26 +11,20 @@ import java.util.stream.Collectors;
 
 @RestController
 public class LineController {
-    private LineDao lineDao;
-    private SectionService sectionService;
     private LineService lineService;
-    private SectionDao sectionDao;
 
-    public LineController(LineDao lineDao, SectionService sectionService, LineService lineService, SectionDao sectionDao) {
-        this.lineDao = lineDao;
-        this.sectionService = sectionService;
+    public LineController(LineService lineService) {
         this.lineService = lineService;
-        this.sectionDao = sectionDao;
     }
 
     @PostMapping("/lines")
     public ResponseEntity<LineResponse> createLine(@RequestBody LineRequest lineRequest) {
         Line line = new Line(lineRequest.getName(), lineRequest.getColor(), lineRequest.getUpStationId(), lineRequest.getDownStationId(), lineRequest.getDistance());
-        try{
+        try {
             Line newLine = lineService.createLine(line);
             LineResponse lineResponse = new LineResponse(newLine.getId(), newLine.getName(), newLine.getColor());
             return ResponseEntity.created(URI.create("/lines/" + newLine.getId())).body(lineResponse);
-        } catch(Exception e){
+        } catch (Exception e) {
             return ResponseEntity.badRequest().build();
         }
     }
@@ -51,15 +40,15 @@ public class LineController {
 
     @GetMapping("/lines/{lineId}")
     public ResponseEntity<LineResponse> showLine(@PathVariable Long lineId) {
-        try{
+        try {
             Line line = lineService.getLineById(lineId);
             List<StationResponse> stationResponses = lineService.getStations(lineId).stream()
                     .map(station -> new StationResponse(station.getId(), station.getName()))
                     .collect(Collectors.toList());
-            System.out.println("stationService.getStationById success");
+
             LineResponse lineResponse = new LineResponse(line.getId(), line.getName(), line.getColor(), stationResponses);
             return ResponseEntity.ok(lineResponse);
-        } catch(Exception e){
+        } catch (Exception e) {
             return ResponseEntity.badRequest().build();
         }
     }
@@ -67,35 +56,14 @@ public class LineController {
     @PutMapping("/lines/{id}")
     public ResponseEntity<LineResponse> updateLine(@PathVariable Long id, @RequestBody LineRequest lineRequest) {
         Line line = new Line(lineRequest.getName(), lineRequest.getColor(), lineRequest.getUpStationId(), lineRequest.getDownStationId());
-        lineDao.updateById(id, line);
+        lineService.updateLineById(id, line);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/lines/{id}")
-    public ResponseEntity deleteStation(@PathVariable Long id) {
-        lineDao.deleteById(id);
+    public ResponseEntity deleteLine(@PathVariable Long id) {
+        lineService.deleteLineById(id);
         return ResponseEntity.noContent().build();
-    }
-
-    @PostMapping("/lines/{lineId}/sections")
-    public ResponseEntity<LineResponse> createSection(@PathVariable Long lineId, @RequestBody SectionRequest sectionRequest) {
-        Section newSection = new Section(lineId, sectionRequest.getUpStationId(), sectionRequest.getDownStationId(), sectionRequest.getDistance());
-        try {
-            sectionService.insert(newSection);
-        } catch (IllegalArgumentException illegalArgumentException) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-        }
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping("/lines/{lineId}/sections")
-    public ResponseEntity<LineResponse> deleteSection(@PathVariable Long lineId, @RequestParam Long stationId) {
-        try {
-            sectionService.delete(lineId, stationId);
-        } catch (IllegalArgumentException illegalArgumentException) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-        }
-        return ResponseEntity.ok().build();
     }
 
 }
